Use Shadow property index when generating Shadow values buffer
Shadow Values factory expects a value buffer which also has slots for shadow navigations. But when populating we did not consider them. So we got a value buffer with values in wrong slots.
It happens only for ModelSnapshot (the shadow model phase) when seeding derived type when base type has a navigation.

It throws NRE if encounter null value for non-null field
or ICE if encounter different type then expected.

Resolves #17851