Revert 3.0 fix for consistent reload and replace with fix that doesn't dramatically change detach behavior
Fixes #17784

In 3.0 we made a [simple change](https://github.com/aspnet/EntityFrameworkCore/commit/7e65b82821ce2540d12d35c6449712f07c56a527#diff-7767de8ee880eb46efbfc9d2153dd33a) for #15645 that started clearing navigation properties when they referenced detached entities. In retrospect this was the wrong thing to do as people are relying on this to be able to detach all entities from the context.

The new fix restores the pre-3.0 detach behavior, and sends the reload of a deleted entity through a simulated to delete instead.