Query: Materialize results of FirstOrDefault in subquery only when there are rows
Issue: When we generate translation for subquery.FirstOrDefault() using Left join, we get null in values when there is no associated record.
Then we try to create result with default values rather than returning default of result type.

Fix: Whenever we go through path where we are adding subquery.FirstOrDefault in projection, we inject 1 as dummy column. If we see 1 in result then only we materialize the object else we return default of the type.
- For entityTypes this is not needed because we return null if key values are null.
- For scalar which is server eval (i.e. only 1 column), we generate projectionBinding using resultType so we will get correct default back.
- For scalar result which is client eval, we inject 1 into client projections.
- For non scalar results (anonymous/nominal types), in client/server eval case, we inject 1 into projection/projectionMapping.

Resolves #17287