Fix to #18725 - Optimize out null checks when translating Enum flags
One of the binary expression optimizations is op(a, b) == null -> a == null || b == null. This is valid for all binary operators except logical (AndAlso, OrElse). Before we used to also skip this optimization for bitwise (And, Or).
However, this optimization is valid for bitwise ops, so the fix is to enable it.

Resolves #18725