Fix to #18808 - Query: incorrect results for query with Join where key selector is anonymous type with a single property
Problem was that on c# join predicate behaves differently when it's a regular property (null and null yield false), and where its anonymous type with just a single property (new { A = null } and new { A = null } yields true).
However, we lose that distinction during initial Join predicate translation.

Fix is to add a sentinel true == true condition, if the anonymous type only has one element. This condition can be later optimized out.

Resolves #18808