Update SQL Server type mappings for strings
Fixes #19327

Create fixed-length type mappings by default, and then use the correct DbType in the parameter if the length of the value matches the parameter length. This has a better chance of hitting an index. In other cases, change the DbType to match what we did before so that SQLClient doesn't truncate or pad.

Fixes #14555

If the parameter value length is greater than the max length but less than the max bounded length, then using the max bounded length rather than going immediately to unbounded.