Query: Simplify SQL generated for optional owned/weak entities when table splitting
Partial fix for #18299

Issue: The way we construct SelectExpression for owned entity type involves adding inner joins to owner when table splitting so that only owned entity for given owner is in result. Considering the fact that any owned entity translation would originate from a query root which would be non-owned, the join condition is unnecessary when it is same as outer query root which initiated this owned entity.
Fix: When we are generating join condition for linking FK if the FK is ownership for entity type and principal type is root entity type then the join condition would be redundant hence we skip adding it.

Notes:
- The joins would be unnecessary even for derived type if the owned navigation is on base type or the given entity type but integrating that information involves breaking change to API. Essentially when generating SelectExpression for owned entity we need to keep track of which selectExpression initiated it.