Cleanup and add documentation for IndentedStringBuilder
Fixes #20383

Looked at making it pubternal or internal again, but it seems to be fundamental to the C# code generators in the design assembly. Considered moving it to design, but that would result in duplicate internal code other places. So just documented and cleaned up the surface--in particular, it only handles strings so that any possible locale-specific string conversions are forced to happen outside.