Obsolete key comparers on properties (#20680)
Part of #20409

Type mappings need to be able to provide different comparers for key and non-key property. However, when a key comparer is being set on a specific property, then a key comparer can be set for key properties and a regular comparer can be set for non-key properties. We don't need to support setting both on the same property.

Also remove the fallback parameter, per API review.