Fix to #20759 - Query: client eval followed by aggregate operation throws KeyNotFoundException: The given key 'EmptyProjectionMember' was not present in the dictionary.
Problem is that when client eval happens in the projection before the aggregate operation SelectExpression projection mapping can/will be empty. However when we try to access them using GetMappedProjection we assume that we always find what we want.
Fix is to check that selectExpression.Projection is empty. Otherwise we know client eval happened and we can't translate aggregate.

Fixes #20759