Query: Make IPrintableExpression.Print explicit interface implementation (#21063)
Issue:
- We have Print extension method on Expression without any additional arguments.
- We also have IPrintableExpression.Print method which takes ExpressionPrinter as argument.

So when calling Print on a arbitrary expression, extension method works fine. But when calling print on a custom expression when type is known (especially in debugger), it asks to pass the ExpressionPrinter argument as it does not pick up extension method.
This is a breaking change but unlikely to cause any issues since Print method on custom expression directly should not be called.