Return true for IsModified on navigations pointing to Added or Deleted entities (#21259)
Fixes #16513

I investigated this some more and I now understand why this was labeled a bug. Collection navigations already have this behavior--that is, Deleted or Added related entities indicate that the navigation property is modified. So this change brings reference navigations inline, which I think is a bug. Nevertheless, I think it makes sense to document this as a breaking change.