Add tests for use GetType in a value converter
Closes #20776

The issue here is external to EF and is a limitation of `Type.GetType()`. GetType can only load types from system assemblies or are resolvable from the calling assembly. This means if GetType is compiled into the materialization expression, then it may not be able to resolve types that it would be able to when called from the application assembly. This can be demonstrated by calling the overload that throws errors:

```
System.TypeLoadException : Could not load type 'Microsoft.EntityFrameworkCore.CustomConvertersTestBase`1+EyeHazTypes' from assembly 'Anonymously Hosted DynamicMethods Assembly, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null'.
   at System.RuntimeTypeHandle.GetTypeByName(String name, Boolean throwOnError, Boolean ignoreCase, StackCrawlMarkHandle stackMark, ObjectHandleOnStack assemblyLoadContext, Boolean loadTypeFromPartialName, ObjectHandleOnStack type, ObjectHandleOnStack keepalive)
   at System.RuntimeTypeHandle.GetTypeByName(String name, Boolean throwOnError, Boolean ignoreCase, StackCrawlMark& stackMark, AssemblyLoadContext assemblyLoadContext, Boolean loadTypeFromPartialName)
   at System.RuntimeType.GetType(String typeName, Boolean throwOnError, Boolean ignoreCase, StackCrawlMark& stackMark)
   at System.Type.GetType(String typeName, Boolean throwOnError)
   at lambda_method789(Closure , QueryContext , DbDataReader , ResultContext , SingleQueryResultCoordinator )
   at Microsoft.EntityFrameworkCore.Query.Internal.SingleQueryingEnumerable`1.Enumerator.MoveNext() in /home/ajcvickers/efcore/src/EFCore.Relational/Query/Internal/SingleQueryingEnumerable.cs:line 157
   at System.Collections.Generic.List`1..ctor(IEnumerable`1 collection)
   at System.Linq.Enumerable.ToList[TSource](IEnumerable`1 source)
   at Microsoft.EntityFrameworkCore.CustomConvertersTestBase`1.Can_query_and_update_with_converter_for_Type_when_GetType_can_load_it() in /home/ajcvickers/efcore/test/EFCore.Specification.Tests/CustomConvertersTestBase.cs:line 133
```

There is no action on the EF side. Calling GetType from the application assembly (as shown) is the correct way to deal with this.