Recategorize the runtime dependencies (#21586)
It is not critically important that these match in exact version to those that are being shipped to NuGet.
They are primarily for verification. During previews, the subscription edge will remain, but once we get to 5.0.0
the edge can be changed to only update on demand.

This ensures that efcore does not end up on the critical path for the build.
EFCore's build is fairly long (though it is currently running tests).