Add SaveChanges events (#21862)
Fixes #15910

I realized that the interceptor, while powerful, is hard to attach to without changing the context configuration. Therefore, this PR adds simple .NET events that can easily be attached to from outside the code that defines the context.

Also, fix to call the SaveChanges failed interceptor when the failure is an optimistic concurrency exception.