Part of #7201
This PR contains the specific changes we made in the meeting, plus using ``{entityType}.{property/navigation}' in EFCore consistently.

The tt files has been updated to allow moving of parameters without changing the method signature and hence without needing to update all call sites, which is error prone.

I have also added other changes we discussed as check boxes in #7201 so that we can track this like we do API reviews.