Mark join entities as Added when either side is Added
Fixes #22647

**Description**

If the entity on either end of a many-to-many relationship is Added, then any join entity instance must be new since it references a new entity via a required relationship. This change improves the many-to-many experience by detecting this case.

**Customer Impact**

This is an improvement to the experience for the new many-to-many feature. Doing this now instead of in the future avoids a breaking change after 5.0. Therefore, doing this now will help customers have a seamless update to 6.0.

**How found**

Customer reported on RC1.

**Test coverage**

Added additional test coverage around this case.

**Regression?**

No; new feature.

**Risk**

Low; small change in behavior.