Handle square brackets around explicitly specified SQL Server type names (#22610)
Fixes #22569

In EF Core 3.1, it was possible to specify a SQL Server column type using square brackets. For example, "[money]" instead of "money". EF Core didn't understand this, so we treated it as an unknown decimal type, and passed it through to SQL Server as-is.

In EF Core 5.0, we added support for precision and scale specified independently of the type. We don't do this when we know the type is `money` because doing so is not valid. However, since we don't recognize "[money]" it gets precision and scale added.

The fix is to recognize SQL Server types with square brackets. Note that only type names without spaces can have square brackets.