[5.0] Fix places where skip navigations are missed
Fixes #22882

**Description**

Skip navigations (i.e. many-to-many navigations; new for 5.0) are not being returned as members, navigations, or collections from an an API used to access them. In addition, scrubbing the code found three other issues with the same pattern, plus a few that need more investigation. This PR contains fixes and tests for the four clear issues. I have filed #22883 to follow up on the others.

**Customer Impact**

* Navigations for a major new feature are not being returned from an API that should return them.
* Attaching a graph asynchronously may miss navigations.
* Skip navigations are missed as notification properties.
* Exception will say a property is not found, when actually it is a navigation.

**How found**

Customer reported on RC2 daily build.

**Test coverage**

This PR adds tests for the affected cases, including others found from scrubbing the code. #22883 tracks doing more research in this area.

**Regression?**

No.

**Risk**

Low. The change only affects skip navigations, which are new in 5.0.