SqlServer Migrations: Don't omit ALTER COLUMN when old type is unknown
The fix for #21364 broke migrations created prior to version 3.1.0 which did not specify the old column type. In this case, we don't know whether the column was previously part of an index or not, so we should always include the ALTER COLUMN statement just in case. Note, the old column type is always specified for migrations created with versions 3.1.0 and newer.

Fixes #23045