Query: Throw error for translating Lambda right away (#23483)
LambdaExpression can never be translated. The only time when returning null has effect in translation is projection where we would visit the base to client eval.
But lambda cannot be client evaluated either so right thing to do is just throw exception so that user know which lambda is throwing error.

Resolves #18179