[5.0.3/Tests only] Updates many-to-many change tracking tests to work with generated key values (#23807)
This is a precursor to fixes for #23659, #23787. This test has no product changes, it just refactors the many-to-many tests so that they can be run with generated key values, as well as running with explicit keys values like they currently do. Generated keys result in more work being done in fixup by navigations, which is where both of these issues live.

Once this is merged I will send out separate PRs to fix the two bugs.