Correctly set pool size when using AddDbContextFactory
Fixes #23889

**Description**

AddPooledDbContextFactory (new in 5.0) has a parameter to set the pool size. This value was not being passed down to lower-level config and hence the default pool size is always used.

**Customer Impact**

The pool size cannot be changed when using the 'AddDbContext' for a factory and pooling.

**How found**

Customer reported on 5.0.

**Test coverage**

The test factoring did not test the defaults or overridden values for pool size; this has been added.

**Regression?**

No; new feature in 5.0.

**Risk**

Very low.