Draft: Mapping a single property to multiple columns
Part of #13947

The idea here is that a property can contain pseudo-properties that generally don't act like properties in the model, but form the basis for mapping to a column after extracting a value from the outer property.

Why properties? Because they have all the metadata needed to allow the facets of columns to be configured?

Why not relational only? Cosmos won't use this, but in-memory will be able to understand it.