Remove inheritance from InternalEntityEntry
I think this is a good change in of itself since the subtypes are even less useful now than when it was originally written, and it allows this type to be sealed.

In addition, this is part of #13947, since the two subclasses will no longer cover the different ways of reading a property value.