Fix to #25511 - Temporal Tables: The given key 'SqlServer:TemporalHistoryTableName' was not present in the dictionary when creating migration on entity mapped to temporal table without explicitly providing history table name
Problem was that we don't store the default history table name in the annotation (like we do with default period property names), instead the getter on IReadOnlyEntityType retrieves the default value if needed.

Fix is to check if the history table annotation is present in the annotation list, just like we do with the history table schema.

Fixes #25511