Query: Clone shaped query expression during translation
Generally we encounter ShapedQueryExpression during translation when translating subqueries
Since ShapedQueryExpression contains QueryExpression which is mutated in-place, we lose the original expression if for any reason translation doesn't work out. Which blocks us from translating it further without recreating original tree again.
We were hitting this during client eval in projection earlier where owned entities are expanded during translation
We recreated tree by expanding owned navigation during the method to retain original expression
With issue #22089 we ran into situation where we try to translate a subquery which is not fully translatable and then we have to visit it through base.
So to avoid the whole issue with reconstruction, now we clone the shaped query during translation so we never mutate original expression

Resolves #22089