Fix to #24953 - Different orders of Include()/Where() in queries produce different results
Problem was that we were not merging includes for the EntityReference that we retrieve from foreign key expansion cache. If the translation was added to the cache before all the includes were marked and more includes were added afterwards, those includes would be lost.
Fix is to merge includes just like we do for the scenario where we need to expand the FK translation.

Fixes #24953