Fix to #24671 - Relational: No tests for TranslateContains path
All Contains calls are converted to Any with predicate (which allows to handle null semantics) so this code never gets but fixing this tiny bug, just in case. We were not setting the cardinality of the contains translation to single result, so the translation would have failed, if it wasn't converted to Any and therefore (correctly) being handled by it's translation code.

Fixes #24671