Fixing bugs around hidden period columns
Fixes #25841 - Create period columns as hidden
Fixes #25774 - EFCore.SqlServer - v6.00-preview - SqlServerDatabaseModelFactory doesn't read information about hidden temoral tables' columns
Fixes #25773 - EFCore.SqlServer - v6.00-preview - failed on creating model for WideWorldImporters sample database temporal tables

Multiple problems were causing the issues in the scaffolding code path:

We used to skip over hidden columns when constructing database model. This made sense when temporal tables were not supported, but now we need that column information.
However, if we add period columns to database model, they will in turn produce EF model with period properties, and then code generator will produce property code for them on the entity, rendering them non-shadow.
All period properties must be in shadow state (otherwise validation fails) and there is currently no way to mark a property as shadow state in the code generator.

Instead, we continue to not add period columns to the table in the database model, so the properties won't be part of the EF model passed to the code generator and therefore code for them is not generated and they de facto stay in shadow state.
To compensate, in SqlServerDatabaseModelFactory we store the period information on the table in form of period start/end property name. Code generator uses annotations to produce the temporal table configuration fluen API, so this information is sufficient.

Also adjusted code in SqlServerAnnotationProvider which was always expecting period properties to always be defined on the model. It is still the case for a normal code path, but for scaffolding properties are missing (due to shadow state code gen issue), and we no longer throw NRE because of that.

Small drawback is that we don't scaffold indexes based on period columns. We would need to add the period columns to the database model in SqlServerDatabaseModelFactory. It's possible to define such index in the model and migrations will create DML for it, but reverse engineer is unable to pick up existing indexes.