Fix to #26014 - Migrations/Temporal Tables: null value exception when executing Down migration to convert entity from temporal to normal
Problem was that we perform a null check on old and new annotation, but null is a valid value for annotation. We should allow this and also mark the annotation value as nullable object. Also fixed small issue in code gen where we would not generate code to add annotation with null value (but we were adding it for old annotation).

Fixes #26014