Add AutoRetry flag for commands
Add an AutoRetry flag to SqliteCommands, making the behavior of
automatically retrying commands when the database is busy or locked
into an optional behavior.  The default is True, so nothing has
changed if the flag is not set to False.

The flag is present at a few levels of the design:

* SqliteCommand has a new AutoRetry property, which takes its value
  from the SqliteConnection by default.  If there isn't a connection
  associated with the command, the default is true.
* SqliteConnection has a new AutoRetry property, which takes its
  value from the connection string by default.
* SqliteConnectionStringBuilder recognizes a new "Auto Retry" keyword
  in connection strings, with true being the default when not
  specified.

The flag is used in the two existing places in the design where
automatic retry behavior was implemented:

* In SqliteDataReader.NextResult().
* In SqliteCommand.PrepareAndEnumerateStatements().

A variety of unit tests are included, verifying aspects of the new
feature's implementation.